% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{budget_execution}
\alias{budget_execution}
\title{Budget Execution Rates}
\format{
A tibble with 2368 rows and 3 columns:
\describe{
  \item{country_code}{ISO country code (character)}
  \item{year}{Year (character)}
  \item{budget_execution_rate}{Budget execution rate. Numeric value representing the percentage of the budget originally approved that was executed.}
}
}
\source{
World Bank. \url{https://data360.worldbank.org/en/indicator/WB_WDI_GF_XPD_BUDG_ZS}
}
\usage{
budget_execution
}
\description{
Annual budget execution rates by country.
}
\keyword{datasets}
