% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_funs.R
\name{calculate_coverage}
\alias{calculate_coverage}
\title{Calculate coverage for an indicator}
\usage{
calculate_coverage(indicator, id)
}
\arguments{
\item{indicator}{A vector containing indicator values. May include `NA`s.}

\item{id}{A vector of identifiers (e.g., country codes, years) corresponding
to the indicator values.}
}
\description{
This function computes the number of distinct identifiers (`id`)
for which a given indicator has non-missing values. It is useful
to measure the coverage of an indicator across countries, years,
or any other grouping identifier.
}
