% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_funs.R
\name{calculate_time_range}
\alias{calculate_time_range}
\title{Calculate the time range of an indicator}
\usage{
calculate_time_range(indicator, time_id)
}
\arguments{
\item{indicator}{A vector of indicator values.}

\item{time_id}{A vector of years corresponding to the indicator values.}
}
\value{
A character string of the form `"start_year-end_year"`.
}
\description{
Returns the first and last year (inclusive) for which an indicator
has non-missing values.
}
