% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d360_efi_data}
\alias{d360_efi_data}
\title{World Bank EFI Data 360 Indicators}
\format{
A data frame with 8878 rows and 189 variables:
\describe{
  \item{country_code}{The World Bank World Bank-3166 country code}
  \item{year}{Year}
  \item{imf_fm_g_x_g01_gdp_pt}{Expenditure (Percent of GDP)}
  \item{imf_fm_g_xwdg_g01_gdp_pt}{Gross debt (Percent of GDP)}
  \item{imf_fm_ggcb_g01_pgdp_pt}{Cyclically adjusted  balance Percent of  GDP}
  \item{imf_fm_ggcbp_g01_pgdp_pt}{Cyclically adjusted primary balance Percent of potential GDP}
  \item{imf_fm_ggr_g01_gdp_pt}{Government revenue, percent of GDP}
  \item{imf_fm_ggxcnl_g01_gdp_pt}{Net lending/borrowing (also referred as overall balance  percent  of GDP}
  \item{imf_fm_ggxonlb_g01_gdp_pt}{Primary net lending/borrowing (also referred as primary balance) (Percent of GDP)}
  \item{imf_fm_ggxwdn_g01_gdp_pt}{Net debt (Percent of GDP)}
  \item{imf_gfscofog_geaf_g14_gdp_pt}{Expenditure on fuel & energy, Percent of GDP}
  \item{imf_gfscofog_ged_g14_gdp_pt}{Expenditure on defense, Percent of GDP}
  \item{imf_gfscofog_gedc_g14_gdp_pt}{Expenditure on civil defense, Percent of GDP}
  \item{imf_gfscofog_gedm_g14_gdp_pt}{Expenditure on military defense, Percent of GDP}
  \item{imf_gfscofog_gednec_g14_gdp_pt}{Expenditure on defense n.e.c., Percent of GDP}
  \item{imf_gfscofog_gedr_g14_gdp_pt}{Expenditure on defense R&D, Percent of GDP}
  \item{imf_gfscofog_gedto_g14_pt}{Expenditure on defense, Percent of total expenditure}
  \item{imf_gfscofog_gee_g14_gdp_pt}{Expenditure on education, Percent of GDP}
  \item{imf_gfscofog_geee_g14_gdp_pt}{Expenditure on subsidiary services to education, Percent of GDP}
  \item{imf_gfscofog_geel_g14_gdp_pt}{Expenditure on education not definable by level, Percent of GDP}
  \item{imf_gfscofog_geen_g14_gdp_pt}{Expenditure on post-secondary non-tertiary education, Percent of GDP}
  \item{imf_gfscofog_geeo_g14_gdp_pt}{Expenditure on education n.e.c., Percent of GDP}
  \item{imf_gfscofog_geep_g14_gdp_pt}{Expenditure on pre-primary & primary education, Percent of GDP}
  \item{imf_gfscofog_geer_g14_gdp_pt}{Expenditure on education R&D, Percent of GDP}
  \item{imf_gfscofog_gees_g14_gdp_pt}{Expenditure on secondary education, Percent of GDP}
  \item{imf_gfscofog_geet_g14_gdp_pt}{Expenditure on tertiary education, Percent of GDP}
  \item{imf_gfscofog_geeto_g14_pt}{Expenditure on education, Percent of total expenditure}
  \item{imf_gfscofog_gegpb_g14_gdp_pt}{Expenditure on basic research, Percent of GDP}
  \item{imf_gfscofog_gegpc_g14_gdp_pt}{Expenditure on exec/leg, fiscal, & external affairs, Percent of GDP}
  \item{imf_gfscofog_gegpt_g14_gdp_pt}{Transfers between different levels of govt, Percent of GDP}
  \item{imf_gfscofog_gehw_g14_gdp_pt}{Expenditure on water supply, Percent of GDP}
  \item{imf_gfscofog_gel_g14_gdp_pt}{Expenditure on health, Percent of GDP}
  \item{imf_gfscofog_gelh_g14_gdp_pt}{Expenditure on hospital services, Percent of GDP}
  \item{imf_gfscofog_gelm_g14_gdp_pt}{Expenditure on medical products, appliances, & equip, Percent of GDP}
  \item{imf_gfscofog_gelp_g14_gdp_pt}{Expenditure on public health services, Percent of GDP}
  \item{imf_gfscofog_gelr_g14_gdp_pt}{Expenditure on health R&D, Percent of GDP}
  \item{imf_gfscofog_gelto_g14_pt}{Expenditure on health, Percent of total expenditure}
  \item{imf_gfscofog_genb_g14_gdp_pt}{Expenditure on biodiversity & landscape protection, Percent of GDP}
  \item{imf_gfscofog_genm_g14_gdp_pt}{Expenditure on waste management, Percent of GDP}
  \item{imf_gfscofog_geno_g14_gdp_pt}{Expenditure on environmental protection n.e.c., Percent of GDP}
  \item{imf_gfscofog_genp_g14_gdp_pt}{Expenditure on pollution abatement, Percent of GDP}
  \item{imf_gfscofog_genr_g14_gdp_pt}{Expenditure on environmental protection R&D, Percent of GDP}
  \item{imf_gfscofog_gento_g14_pt}{Expenditure on environment protection, Percent of total expenditure}
  \item{imf_gfscofog_genw_g14_gdp_pt}{Expenditure on waste water management, Percent of GDP}
  \item{imf_gfscofog_gep_g14_gdp_pt}{Expenditure on public order & safety, Percent of GDP}
  \item{imf_gfscofog_gepf_g14_gdp_pt}{Expenditure on fire protection services, Percent of GDP}
  \item{imf_gfscofog_gepl_g14_gdp_pt}{Expenditure on law courts, Percent of GDP}
  \item{imf_gfscofog_gepo_g14_gdp_pt}{Expenditure on public order & safety n.e.c., Percent of GDP}
  \item{imf_gfscofog_gepp_g14_gdp_pt}{Expenditure on prWorld Bankns, Percent of GDP}
  \item{imf_gfscofog_gepr_g14_gdp_pt}{Expenditure on public order & safety R&D, Percent of GDP}
  \item{imf_gfscofog_geps_g14_gdp_pt}{Expenditure on police services, Percent of GDP}
  \item{imf_gfscofog_gepto_g14_pt}{Expenditure on public order & safety, Percent of total expenditure}
  \item{imf_gfscofog_gesp_g14_gdp_pt}{Expenditure on social protection, Percent of GDP}
  \item{imf_gfscofog_gespp_g14_gdp_pt}{Expenditure on social protection n.e.c., Percent of GDP}
  \item{imf_gfscofog_gespr_g14_gdp_pt}{Expenditure on social protection R&D, Percent of GDP}
  \item{imf_gfscofog_gesps_g14_gdp_pt}{Expenditure on sickness  & disability, Percent of GDP}
  \item{imf_gfscofog_gespto_g14_pt}{Expenditure on social protection, Percent of total expenditure}
  \item{imf_gfscofog_gespu_g14_gdp_pt}{Expenditure on unemployment, Percent of GDP}
  \item{imf_gfscofog_geto_g14_gdp_pt}{Expenditure, Percent of GDP}
  \item{wb_wdi_iq_sci_mthd}{Methodology assessment of statistical capacity (scale 0 - 100)}
  \item{wb_wdi_iq_sci_prdc}{Periodicity and timeliness assessment of statistical capacity (scale 0 - 100)}
  \item{wb_lpi_lp_lpi_cust_xq}{Logistics Performance Index: Efficiency of customs clearance processes (i.e. speed, simplicity and predictability of formalities) by border control agencies, including customs (scale 1-5)}
  \item{wb_wdi_si_pov_mdim}{The percentage of people who are multidimensionally poor}
  \item{wb_wdi_si_pov_mdim_it}{Average share of deprivations experienced by the poor}
  \item{wb_wdi_si_pov_mdim_xq}{Multidimensional poverty index, children (population ages 0-17) (scale 0-1)}
  \item{wb_wwbi_bi_emp_frml_pb_ed_zs}{Public sector employment, as a share of formal employment, by industry: Education}
  \item{wb_wwbi_bi_emp_frml_pb_he_zs}{Public sector employment, as a share of formal employment, by industry: Health}
  \item{wb_wwbi_bi_emp_frml_pb_zs}{Measures the number of public sector paid employees}
  \item{wb_wwbi_bi_emp_pwrk_ed_pb_zs}{Education workers, as a share of public paid employees}
  \item{wb_wwbi_bi_emp_pwrk_he_pb_zs}{Health workers, as a share of public paid employees}
  \item{wb_wwbi_bi_emp_totl_no}{Number of employed individuals}
  \item{wb_wwbi_bi_emp_totl_no_ed}{Number of employed employees, by industry: Education}
  \item{wb_wwbi_bi_emp_totl_no_he}{Number of employed employees, by industry: Health}
  \item{wb_wwbi_bi_emp_totl_no_pa}{Number of employed employees, by industry: Public adminstration}
  \item{wb_wwbi_bi_emp_totl_pb_tt_zs}{Proportion of total employees with tertiary education working in public sector}
  \item{wb_wwbi_bi_pwk_prvs_tt_zs}{Individuals with tertiary education as a share of private paid employees}
  \item{wb_wwbi_bi_pwk_pubs_fe_zs}{Number of female public paid employees/Total number of public paid employees.}
  \item{wb_wwbi_bi_pwk_pubs_no}{Number of public paid employees}
  \item{wb_wwbi_bi_pwk_pubs_no_ed}{Number of public paid employees, by industry: Education}
  \item{wb_wwbi_bi_pwk_pubs_no_he}{Number of public paid employees, by industry: Health}
  \item{wb_wwbi_bi_pwk_pubs_no_pa}{Number of public paid employees, by industry: Public adminstration}
  \item{wb_wwbi_bi_pwk_pubs_sn_fe_zs}{Number of female public paid employees who work as senior official (managers) / Number of public paid employees work as senior official (managers).}
  \item{wb_wwbi_bi_pwk_pubs_tt_zs}{Measures the number of public paid employees with tertiary education as a share of total public employment}
  \item{wb_wwbi_bi_pwk_totl_no}{Number of paid employees}
  \item{wb_wwbi_bi_pwk_totl_no_ed}{Number of paid employees, by industry: Education}
  \item{wb_wwbi_bi_pwk_totl_no_he}{Number of paid employees, by industry: Health}
  \item{wb_wwbi_bi_pwk_totl_no_pa}{Number of paid employees, by industry: Public adminstration}
  \item{wb_wwbi_bi_wag_cprs_pb_zs}{Ratios of the 90th percentile of weekly wage for public paid employees/10th percentile of weekly wage for public paid employees.}
  \item{wb_wwbi_bi_wag_cprs_pv_zs}{Pay compression ratio in private sector (ratio of 90th/10th percentile earners)}
  \item{wb_wwbi_bi_wag_prem_ed_gp}{Public sector wage premium, by industry: Education (compared to all private employees)}
  \item{wb_wwbi_bi_wag_prem_he_gp}{Public sector wage premium, by industry: Health (compared to all private employees)}
  \item{wb_wwbi_bi_wag_prem_pb_gp}{Percentage differences in public sector wages compared to private sector wages (in local currency units) controlling for education, age, gender, and location.}
  \item{wb_wwbi_bi_wag_totl_gd_zs}{General government wave bill in proportion to country GDP (based on PPP; 2009 dollars). The wage bill is defined as the total compensation (in cash or in-kind) payable to a government employee in exchange for work. Wage bill includes wages and salaries, allowances, and social security contributions made on behalf of employees to social insurance schemes.}
  \item{wb_wwbi_bi_wag_totl_pb_zs}{Measures the total compensation (in cash or in-kind) payable to a government employee in exchange for work.}
  \item{bs_bti_q1_2}{Measures whether all relevant groups in society agree about citizenship and accept the nation-state as legitimate}
  \item{bs_bti_q12_1}{Measures the extent to which environmental concerns effectively taken into account}
  \item{bs_bti_q12_2}{To what extent are there solid institutions for basic, secondary and tertiary education, as well as for research and development?}
  \item{bs_bti_q14}{Measures whether the government manages reforms effectively and can achieve its policy priorities.}
  \item{bs_bti_q14_1}{To what extent does the government set and maintain strategic priorities?}
  \item{bs_bti_q14_2}{How effective is the government in implementing its own policies?}
  \item{bs_bti_q14_3}{How innovative and flexible is the government?}
  \item{bs_bti_q15_1}{Measures the extent to which the government makes efficient use of available human, financial and organizational resources}
  \item{bs_bti_q15_2}{To what extent can the government coordinate conflicting objectives into a coherent policy?}
  \item{bs_bti_q2_1}{Captures the extent to which elections are free and fair, considering factors such as suffrage, the secret ballot, electoral management, voter registration, media access, etc. (1 to 10 scoring)}
  \item{bs_bti_q2_3}{Captures freedom of association or the extent to which individuals can freely form and join independent political or civic groups}
  \item{bs_bti_q3_1}{Captures the extent to which there is a working separation of powers (i.e. checks and balances)}
  \item{bs_bti_q3_2}{Measures the independence of the judiciary (i.e. the ability and autonomy to interpret and review existing law, pursue its own reasoning free from the influence of political decision-makers or powerful groups and individual, etc)}
  \item{bs_bti_q7_2}{Captures  if free and fair competition is guaranteed by an institutional framework that ensures unrestricted participation in the market and a level playing field for all market participants.}
  \item{bs_bti_q7_4}{To what extent have a solid banking system and a functioning capital market been established?}
  \item{bs_bti_q8_1}{Measures the extent to which the monetary authority pursues and communicates a consistent monetary stabilization policy}
  \item{bs_bti_q8_2}{Measures the extent t which the government’s budgetary policies support fiscal stability}
  \item{bs_bti_si}{Composed of five underlying criteria: stateness, political participation, rule of law, stability of democratic institutions, and political and social integration}
  \item{imf_world_rg_rm_gdp}{Total Grants in Percent of GDP}
  \item{imf_world_rs_rm_gdp}{Social Contributions in Percent of GDP}
  \item{imf_world_rt_rm_gdp}{Tax Revenue in Percent of GDP}
  \item{imf_world_rtgse_rm_gdp}{Excise Tax Revenue in Percent of GDP}
  \item{imf_world_rtgsgv_rm_gdp}{VAT Revenue in Percent of GDP}
  \item{imf_world_rti_rm_gdp}{Income Tax Revenue in Percent of GDP}
  \item{imf_world_rtic_rm_gdp}{Corporate Income Tax Revenue in Percent of GDP}
  \item{imf_world_rtii_rm_gdp}{Individual Income Tax Revenue in Percent of GDP}
  \item{imf_world_rtp_rm_gdp}{Property Tax Revenue in Percent of GDP}
  \item{imf_world_rtpay_rm_gdp}{Taxes on Payroll and Workforce Revenue in Percent of GDP}
  \item{imf_world_rtt_rm_gdp}{Tax revenue (Percent of GDP)}
  \item{wb_girg_6}{Captures how policymakers interact with stakeholders when shaping regulations affecting business communities}
  \item{wb_gtmi_cgsi}{Captures the key aspects of a whole-of-government approach, including government cloud, interoperability framework and other platforms.}
  \item{wb_gtmi_dcei}{Measures aspects of public participation platforms, citizen feedback mechanisms, open data, and open government portals.}
  \item{wb_gtmi_gtei}{Measures the strategy, institutions, laws, and regulations to foster GovTech as well as digital skills and innovation policies}
  \item{wb_gtmi_i_12}{Captures whether there is an e-Procurement portal in place}
  \item{wb_gtmi_i_13}{Captures whether there a Debt Management System (DMS) in place}
  \item{wb_gtmi_i_14}{Captures whether there a Public Investment Management System (PIMS) in place}
  \item{wb_gtmi_i_5}{Captures whether there is an operational FMIS in place to support core PFM functions}
  \item{wb_gtmi_i_6}{Captures whether there is a TSA supported by FMIS to automate payments and bank reconciliation}
  \item{wb_gtmi_i_7}{Captures whether there a Tax Management Information System in place}
  \item{wb_gtmi_i_8}{Captures whether there is a Customs Management Information System is in place}
  \item{wb_gtmi_psdi}{Measures the maturity of online public service portals with a focus on citizen centric design and universal accessibility.}
  \item{wb_wdi_dt_dod_mwbg_cd}{IBRD loans and IDA credits (DOD, current US$)}
  \item{wb_wdi_dt_nfl_pcbo_cd}{Commercial banks and other lending (PPG + PNG) (NFL, current US$)}
  \item{wb_wdi_gc_rev_xgrt_gd_zs}{Measures cash receipts from taxes, social contributions, and other revenues such as fines, fees, rent, and income from property or sales}
  \item{wb_wdi_gc_tax_expt_zs}{Taxes on exports (Percent of tax revenue)}
  \item{wb_wdi_gc_tax_gsrv_rv_zs}{Taxes on goods and services (Percent value added of industry and services)}
  \item{wb_wdi_gc_tax_impt_zs}{Customs and other import duties (Percent of tax revenue)}
  \item{wb_wdi_gc_tax_intt_rv_zs}{Taxes on international trade (Percent of revenue)}
  \item{wb_wdi_gc_xpn_comp_zs}{Compensation of employees (Percent of expense)}
  \item{wb_wdi_gc_xpn_gsrv_zs}{Goods and services expense (Percent of expense)}
  \item{wb_wdi_gc_xpn_trft_zs}{Subsidies and other transfers (Percent of expense)}
  \item{wb_wdi_ic_bus_ndns_zs}{New business density (new registrations per 1,000 people ages 15-64)}
  \item{wb_wdi_ic_bus_nreg}{New businesses registered (number)}
  \item{wb_wdi_ie_ppi_watr_cd}{Investment in water and sanitation with private participation (current US$)}
  \item{wb_wdi_iq_spi_ovrl}{Statistical performance indicators (SPI): Overall score (scale 0-100)}
  \item{wb_wdi_iq_spi_pil1}{Statistical performance indicators (SPI): Pillar 1 data use score (scale 0-100)}
  \item{wb_wdi_iq_spi_pil2}{Statistical performance indicators (SPI): Pillar 2 data services score (scale 0-100)}
  \item{wb_wdi_iq_spi_pil3}{Statistical performance indicators (SPI): Pillar 3 data products score (scale 0-100)}
  \item{wb_wdi_iq_spi_pil4}{Statistical performance indicators (SPI): Pillar 4 data sources score (scale 0-100)}
  \item{wb_wdi_iq_spi_pil5}{Statistical performance indicators (SPI): Pillar 5 data infrastructure score (scale 0-100)}
  \item{wb_wdi_tm_val_fuel_zs_un}{Fuel imports (Percent of merchandise imports)}
  \item{wb_gtmi_pfm_mis}{Measures whether there is an operational FMIS, a TSA, a Tax MIS, a Customs MIS, a DMS, a PIMS, and e-procurement in place}
}
}
\source{
World Bank EFI Data 360 (as provided by you)
}
\usage{
d360_efi_data
}
\description{
This dataset contains a cleaned slice of the World Bank EFI (Data 360) database.
}
\details{
Below are the longer descriptions for each variable (when provided):
\itemize{
  \item \strong{imf_fm_g_x_g01_gdp_pt}: Government expenditure, percent of GDP.General government final consumption expenditure (formerly general government consumption) includes all government current expenditures for purchases of goods and services (including compensation of employees). It also includes most expenditures on national defense and security, but excludes government military expenditures that are part of government capital formation.
  \item \strong{imf_fm_g_xwdg_g01_gdp_pt}: WEO: Gross debt consists of all liabilities that require payment or payments of interest and/or principal by the debtor to the creditor at a date or dates in the future. This includes debt liabilities in the form of SDRs, currency and deposits, debt securities, loans, insurance, pensions and standardized guarantee schemes, and other accounts payable. Thus, all liabilities in the GFSM 2001 system are debt, except for equity and investment fund shares and financial derivatives and employee stock options. Debt can be valued at current market, nominal, or face values
  \item \strong{imf_fm_ggcb_g01_pgdp_pt}: Cyclically adjusted  balance Percent of  GDP
  \item \strong{imf_fm_ggcbp_g01_pgdp_pt}: Cyclically adjusted balance excluding net interest payment (interest expenditure minus interest revenue).
  \item \strong{imf_fm_ggr_g01_gdp_pt}: Government revenue, percent of GDP.Tax revenue refers to compulsory transfers to the central government for public purposes. Certain compulsory transfers such as fines, penalties, and most social security contributions are excluded. Refunds and corrections of erroneously collected tax revenue are treated as negative revenue.
  \item \strong{imf_fm_ggxcnl_g01_gdp_pt}: Net lending/borrowing (also referred as overall balance  percent  of GDP
  \item \strong{imf_fm_ggxonlb_g01_gdp_pt}: Overall balance excluding net interest payment (interest expenditure minus interest revenue).
  \item \strong{imf_fm_ggxwdn_g01_gdp_pt}: Gross debt minus financial assets corresponding to debt instruments. These financial assets are: monetary gold and SDRs, currency and deposits, debt securities, loans, insurance, pension, and standardized guarantee schemes, and other accounts receivable. In some countries the reported net debt can deviate from this definition on the basis of available information and national fiscal accounting practices.
  \item \strong{imf_gfscofog_geaf_g14_gdp_pt}: Expenditure on fuel & energy, Percent of GDP
  \item \strong{imf_gfscofog_ged_g14_gdp_pt}: Military expenditure (Percent of GDP).Military expenditures data from SIPRI are derived from the NATO definition, which includes all current and capital expenditures on the armed forces, including peacekeeping forces; defense ministries and other government agencies engaged in defense projects; paramilitary forces, if these are judged to be trained and equipped for military operations; and military space activities. Such expenditures include military and civil personnel, including retirement pensions of military personnel and social services for personnel; operation and maintenance; procurement; military research and development; and military aid (in the military expenditures of the donor country). Excluded are civil defense and current expenditures for previous military activities, such as for veterans' benefits, demobilization, conversion, and destruction of weapons. This definition cannot be applied for all countries, however, since that would require much more detailed information than is available about what is included in military budgets and off-budget military expenditure items. (For example, military budgets might or might not cover civil defense, reserves and auxiliary forces, police and paramilitary forces, dual-purpose forces such as military and civilian police, military grants in kind, pensions for military personnel, and social security contributions paid by one part of government to another.)
  \item \strong{imf_gfscofog_gedc_g14_gdp_pt}: Expenditure on civil defense, Percent of GDP. Civil defense Expenditure are part of the Government expenditure, percent of GDP
  \item \strong{imf_gfscofog_gedm_g14_gdp_pt}: Military expenditure percent of GDP): Military expenditures data from SIPRI are derived from the NATO definition, which includes all current and capital expenditures on the armed forces, including peacekeeping forces; defense ministries and other government agencies engaged in defense projects; paramilitary forces, if these are judged to be trained and equipped for military operations; and military space activities. Such expenditures include military and civil personnel, including retirement pensions of military personnel and social services for personnel; operation and maintenance; procurement; military research and development; and military aid (in the military expenditures of the donor country). Excluded are civil defense and current expenditures for previous military activities, such as for veterans' benefits, demobilization, conversion, and destruction of weapons. This definition cannot be applied for all countries, however, since that would require much more detailed information than is available about what is included in military budgets and off-budget military expenditure items. (For example, military budgets might or might not cover civil defense, reserves and auxiliary forces, police and paramilitary forces, dual-purpose forces such as military and civilian police, military grants in kind, pensions for military personnel, and social security contributions paid by one part of government to another.)
  \item \strong{imf_gfscofog_gednec_g14_gdp_pt}: Expenditure on police services, Percent of GDP
  \item \strong{imf_gfscofog_gedr_g14_gdp_pt}: Government defense R&D funding as a share of GDP
  \item \strong{imf_gfscofog_gedto_g14_pt}: Military expenditure as a percentage of GDP is calculated as the unweighted country average within each country group.
  \item \strong{imf_gfscofog_gee_g14_gdp_pt}: Government expenditure on education, total (percent of GDP) . Expenditure on education are part of the Government expenditure, percent of GDP
  \item \strong{imf_gfscofog_geee_g14_gdp_pt}: Expenditure on subsidiary services to education, Percent of GDP
  \item \strong{imf_gfscofog_geel_g14_gdp_pt}: Expenditure on education not definable by level, Percent of GDP
  \item \strong{imf_gfscofog_geen_g14_gdp_pt}: Expenditure on secondary and primary education (Percent of government expenditure on education)
  \item \strong{imf_gfscofog_geeo_g14_gdp_pt}: Gross domestic expenditures on research and development (R&D), expressed as a percent of GDP. They include both capital and current expenditures in the four main sectors: Business enterprise, Government, Higher education and Private non-profit. R&D covers basic research, applied research, and experimental development.
  \item \strong{imf_gfscofog_geep_g14_gdp_pt}: Expenditure on pre-primary & primary education, Percent of GDP
  \item \strong{imf_gfscofog_geer_g14_gdp_pt}: Public Expenditure on education R&D, Percent of GDP.Gross domestic expenditures on research and development (R&D), expressed as a percent of GDP. They include both capital and current expenditures in the four main sectors: Business enterprise, Government, Higher education and Private non-profit. R&D covers basic research, applied research, and experimental development.
  \item \strong{imf_gfscofog_gees_g14_gdp_pt}: Government expenditure on education, total (Percent of GDP)
  \item \strong{imf_gfscofog_geet_g14_gdp_pt}: Expenditure on tertiary education (Percent of government expenditure on education)
  \item \strong{imf_gfscofog_geeto_g14_pt}: Current health expenditure (Percent of GDP). General government expenditure on education (current, capital, and transfers) is expressed as a percentage of GDP. It includes expenditure funded by transfers from international sources to government. General government usually refers to local, regional and central governments.
  \item \strong{imf_gfscofog_gegpb_g14_gdp_pt}: Expenditure on basic research, Percent of GDP
  \item \strong{imf_gfscofog_gegpc_g14_gdp_pt}: Expenditure on exec/leg, fiscal, & external affairs, Percent of GDP
  \item \strong{imf_gfscofog_gegpt_g14_gdp_pt}: Transfers between different levels of govt, Percent of GDP
  \item \strong{imf_gfscofog_gehw_g14_gdp_pt}: Expenditure on water supply, Percent of GDP
  \item \strong{imf_gfscofog_gel_g14_gdp_pt}: Expenditure on health, Percent of GDP. Level of current health expenditure expressed as a percentage of GDP.  Estimates of current health expenditures include healthcare goods and services consumed during each year. This indicator does not include capital health expenditures such as buildings, machinery, IT and stocks of vaccines for emergency or outbreaks.
  \item \strong{imf_gfscofog_gelh_g14_gdp_pt}: Gross domestic expenditures on research and development (R&D), expressed as a percent of GDP. They include both capital and current expenditures in the four main sectors: Business enterprise, Government, Higher education and Private non-profit. R&D covers basic research, applied research, and experimental development.
  \item \strong{imf_gfscofog_gelm_g14_gdp_pt}: Expenditure on medical products, appliances, & equip (Percent of government expenditure on education)
  \item \strong{imf_gfscofog_gelp_g14_gdp_pt}: Expenditure on public health services, Percent of GDP.Level of current health expenditure expressed as a percentage of GDP.  Estimates of current health expenditures include healthcare goods and services consumed during each year. This indicator does not include capital health expenditures such as buildings, machinery, IT and stocks of vaccines for emergency or outbreaks.
  \item \strong{imf_gfscofog_gelr_g14_gdp_pt}: Expenditure on health R&D, Percent of GDP
  \item \strong{imf_gfscofog_gelto_g14_pt}: Total expenditure on health as a percentage of gross domestic product
  \item \strong{imf_gfscofog_genb_g14_gdp_pt}: Gross domestic expenditures on research and development (R&D), expressed as a percent of GDP  on biodiversity & landscape protection
  \item \strong{imf_gfscofog_genm_g14_gdp_pt}: Expenditure on waste management, Percent of GDP
  \item \strong{imf_gfscofog_geno_g14_gdp_pt}: Expenditure on environmental protection n.e.c., Percent of GDP
  \item \strong{imf_gfscofog_genp_g14_gdp_pt}: Expenditure on pollution abatement , Percent of GDP . International Monetary Fund (IMF), Statistics Department. 2021. Government Finance Statistics (GFS) Database. https://data.imf.org/?sk=a0867067-d23c-4ebc-ad23-d3b015045405. Accessed on 2023-06-17; International Monetary Fund (IMF), Statistics Department (Government Finance Division) Questionnaire.
  \item \strong{imf_gfscofog_genr_g14_gdp_pt}: Expenditure on environmental protection R&D as percent of GDP
  \item \strong{imf_gfscofog_gento_g14_pt}: Government spends on environmental protection measures, as a percentage of the country’s GDP. These measures are part of a specified set of activities, as outlined by the framework of the Classification of Functions of Government (COFOG), and include pollution abatement, protection of biodiversity, waste management and more.
  \item \strong{imf_gfscofog_genw_g14_gdp_pt}: Expenditure on waste water management, Percent of GDP
  \item \strong{imf_gfscofog_gep_g14_gdp_pt}: Expenditure on public order & safety, Percent of GDP
  \item \strong{imf_gfscofog_gepf_g14_gdp_pt}: Expenditure on fire protection services (Percent of government expenditure on civil defense spending)
  \item \strong{imf_gfscofog_gepl_g14_gdp_pt}: Expenditure on law courts, Percent of GDP
  \item \strong{imf_gfscofog_gepo_g14_gdp_pt}: Expenditure on public order & safety n.e.c., Percent of GDP
  \item \strong{imf_gfscofog_gepp_g14_gdp_pt}: Expenditure on prWorld Bankns, Percent of GDP
  \item \strong{imf_gfscofog_gepr_g14_gdp_pt}: Government current expenditures: Public order and safety  R&D, Percent of GDP
  \item \strong{imf_gfscofog_geps_g14_gdp_pt}: Expenditure on police services, Percent of GDP
  \item \strong{imf_gfscofog_gepto_g14_pt}: Government expenditure on public order and safety, percent of total expenditure
  \item \strong{imf_gfscofog_gesp_g14_gdp_pt}: Expenditure on social protection, Percent of GDP.  Social spending is measured on the basis of annual data on government spending on health and education using a database created by the Fiscal Affairs Department (FAD), and checked for accuracy by IMF staff from each country desk.  Source: Social Spending and Social Protection in IMF-Supported Programs
  \item \strong{imf_gfscofog_gespp_g14_gdp_pt}: Government expenditures by function of social protection as percentage of GDP
  \item \strong{imf_gfscofog_gespr_g14_gdp_pt}: Public Expenditure social_x000D_  protection as_x000D_  Percent of GDP
  \item \strong{imf_gfscofog_gesps_g14_gdp_pt}: Expenditure on sickness  & disability (Percent of government expenditure on social spending)
  \item \strong{imf_gfscofog_gespto_g14_pt}: Social expenditure comprises cash benefits, direct in-kind provision of goods and services, and tax breaks with social purposes. Benefits may be targeted at low-income households, the elderly, disabled, sick, unemployed, or young persons. To be considered "social", programmes have to involve either redistribution of resources across households or compulsory participation. Social benefits are classified as public when general government (that is central, state, and local governments, including social security funds) controls the relevant financial flows. All social benefits not provided by general government are considered private. Private transfers between households are not considered as "social" and not included here. Net total social expenditure includes both public and private expenditure. It also accounts for the effect of the tax system by direct and indirect taxation and by tax breaks for social purposes. This indicator is measured as a percentage of GDP or USD per capita.
  \item \strong{imf_gfscofog_gespu_g14_gdp_pt}: Public unemployment spending total, percent of GDP. Unemployment spending  are sub-part of the Government expenditure, percent of GDP
  \item \strong{imf_gfscofog_geto_g14_gdp_pt}: Government expenditure, percent of GDP Percent of GDP. Source. https://www.imf.org/external/datamapper/exp@FPP/USA/FRA/JPN/GBR/SWE/ESP/ITA/ZAF/IND
  \item \strong{wb_wdi_iq_sci_mthd}: The methodology indicator measures a country’s ability to adhere to internationally recommended standards and methods. The methodology score is calculated as the weighted average of 10 underlying indicator scores. The final methodology score contributes 1/3 of the overall Statistical Capacity Indicator score
  \item \strong{wb_wdi_iq_sci_prdc}: The periodicity and timeliness indicator assesses the availability and periodicity of key socioeconomic indicators. It measures the extent to which data are made accessible to users through transformation of source data into timely statistical outputs. The periodicity score is calculated as the weighted average of 10 underlying indicator scores. The final periodicity score contributes 1/3 of the overall Statistical Capacity Indicator score.
  \item \strong{wb_wdi_si_pov_mdim}: The percentage of people who are multidimensionally poor
  \item \strong{wb_wdi_si_pov_mdim_it}: The average percentage of dimensions in which poor people are deprived
  \item \strong{wb_wdi_si_pov_mdim_xq}: Proportion of the child population that is multidimensionally poor adjusted by the intensity of the deprivations
  \item \strong{wb_wwbi_bi_emp_frml_pb_ed_zs}: Public sector employment, as a share of formal employment, by industry: Education
  \item \strong{wb_wwbi_bi_emp_frml_pb_he_zs}: Public sector employment, as a share of formal employment, by industry: Health
  \item \strong{wb_wwbi_bi_emp_frml_pb_zs}: Number of public sector paid employees/ Number of formal employees. Formal employment is defined by having access to at least one of following benefits (contract, health insurance, union membership, or social security)
  \item \strong{wb_wwbi_bi_emp_pwrk_ed_pb_zs}: Education workers, as a share of public paid employees
  \item \strong{wb_wwbi_bi_emp_pwrk_he_pb_zs}: Health workers, as a share of public paid employees
  \item \strong{wb_wwbi_bi_emp_totl_no}: Number of employed individuals
  \item \strong{wb_wwbi_bi_emp_totl_no_ed}: Number of employed employees, by industry: Education
  \item \strong{wb_wwbi_bi_emp_totl_no_he}: Number of employed employees, by industry: Health
  \item \strong{wb_wwbi_bi_emp_totl_no_pa}: Number of employed employees, by industry: Public adminstration
  \item \strong{wb_wwbi_bi_emp_totl_pb_tt_zs}: Proportion of total employees with tertiary education working in public sector
  \item \strong{wb_wwbi_bi_pwk_prvs_tt_zs}: Individuals with tertiary education as a share of private paid employees
  \item \strong{wb_wwbi_bi_pwk_pubs_fe_zs}: Number of female public paid employees/Total number of public paid employees. The indicator here is based on paid employee only. Thus, we exclude working individual with other employment type (selfemployed, non-paid employee, employer, etc.) when calculating this indicator. It should not be viewed as the share of female in labor force.
  \item \strong{wb_wwbi_bi_pwk_pubs_no}: Number of public paid employees
  \item \strong{wb_wwbi_bi_pwk_pubs_no_ed}: Number of public paid employees, by industry: Education
  \item \strong{wb_wwbi_bi_pwk_pubs_no_he}: Number of public paid employees, by industry: Health
  \item \strong{wb_wwbi_bi_pwk_pubs_no_pa}: Number of public paid employees, by industry: Public adminstration
  \item \strong{wb_wwbi_bi_pwk_pubs_sn_fe_zs}: Number of female public paid employees who work as senior official (managers) / Number of public paid employees work as senior official (managers). Managers plan, direct, coordinate and evaluate the overall activities of enterprises, governments, and other organizations. This classification includes chief executives, senior officials, legislators,_x000D_ and managers of any kind.
  \item \strong{wb_wwbi_bi_pwk_pubs_tt_zs}: Number of public (private) paid employees with tertiary education (no education, primary education, secondary education / Total number of public (private) paid employees
  \item \strong{wb_wwbi_bi_pwk_totl_no}: Number of paid employees
  \item \strong{wb_wwbi_bi_pwk_totl_no_ed}: Number of paid employees, by industry: Education
  \item \strong{wb_wwbi_bi_pwk_totl_no_he}: Number of paid employees, by industry: Health
  \item \strong{wb_wwbi_bi_pwk_totl_no_pa}: Number of paid employees, by industry: Public adminstration
  \item \strong{wb_wwbi_bi_wag_cprs_pb_zs}: Ratios of the 90th percentile of weekly wage for public paid employees/10th percentile of weekly wage for public paid employees.
  \item \strong{wb_wwbi_bi_wag_cprs_pv_zs}: Pay compression ratio in private sector (ratio of 90th/10th percentile earners)
  \item \strong{wb_wwbi_bi_wag_prem_ed_gp}: Public sector wage premium, by industry: Education (compared to all private employees)
  \item \strong{wb_wwbi_bi_wag_prem_he_gp}: Public sector wage premium, by industry: Health (compared to all private employees)
  \item \strong{wb_wwbi_bi_wag_prem_pb_gp}: Percentage differences in public sector wages compared to private sector wages (in local currency units) controlling for education, age, gender, and location.
  \item \strong{wb_wwbi_bi_wag_totl_gd_zs}: General government wave bill in proportion to country GDP (based on PPP; 2009 dollars). The wage bill is defined as the total compensation (in cash or in-kind) payable to a government employee in exchange for work. Wage bill includes wages and salaries, allowances, and social security contributions made on behalf of employees to social insurance schemes.
  \item \strong{wb_wwbi_bi_wag_totl_pb_zs}: The wage bill is defined as the total compensation (in cash or in-kind) payable to a government employee in exchange for work. Wage bill includes wages and salaries, allowances, and social security contributions made on behalf of employees to social insurance schemes.
  \item \strong{bs_bti_q1_2}: To what extent do all relevant groups in society agree about citizenship and accept the nation-state as legitimate?
  \item \strong{bs_bti_q12_1}: To what extent are environmental concerns effectively taken into account?
  \item \strong{bs_bti_q12_2}: This question seeks to assess:_x000D_ · whether education policy is successful in delivering high-quality education and training_x000D_ · whether research and development receive effective support from the government_x000D_ When answering this question, your focus should not be on expenditures alone, but also on the quality and competitiveness_x000D_ of the education system and the research sector. Please also consider:_x000D_ · the structure of funding and knowledge providers (public, private, international cooperation)_x000D_ · the output of the educational and developmental efforts: enrollment rates, literacy rates, percentage of people_x000D_ with higher education, number of patent applications etc._x000D_ Please note that unequal access to education should be primarily discussed in indicator 10.2 (Equal Opportunity)._x000D_ Quantitative Reference Indicators: UN Education Index · Public expenditure on education · R&D expenditure
  \item \strong{bs_bti_q14}: Measures whether the government manages reforms effectively and can achieve its policy priorities. Index based on Prioritization (To what extent does the government set and maintain strategic priorities?), Implementation (How effective is the government in implementing its own policies?) and Policy Learning (How innovative and flexible is the government?).
  \item \strong{bs_bti_q14_1}: This question seeks to assess:_x000D_ · the political capability to take on a longer-term perspective going beyond immediate concerns of electoral_x000D_ competition and to maintain strategic priorities over periods of crisis and stalemate_x000D_ · the strategic capacity of the government to prioritize and organize its policy measures (gaining and organizing_x000D_ expertise, evidence-based policymaking, regulatory impact assessments, strategic planning units)_x000D_ The focus should be on the executive, including the administration and the cabinet. Make sure to identify reform_x000D_ drivers and defenders of the status quo, as political determination and institutional capacity may vary among different_x000D_ departments and ministries. Please also comment on how setting and maintaining strategic priorities might be_x000D_ constrained by government composition and by actors outside the government (e.g. powerful economic interests,_x000D_ lobbies, foreign governments, foreign donors)._x000D_ Please consider if the strategic priorities correspond with the normative framework of the BTI in terms of striving for_x000D_ democracy and a market economy. The maximum score for autocracies is 5 points.
  \item \strong{bs_bti_q14_2}: This question asks whether the political leadership involves civil society actors in:_x000D_ · agenda setting_x000D_ · policy formulation_x000D_ · deliberation and decision-making_x000D_ · policy implementation_x000D_ · performance monitoring_x000D_ Civil society actors include civic, economic and professional interest associations, religious, charity and communitybased_x000D_ organizations, intellectuals, scientists and journalists.
  \item \strong{bs_bti_q14_3}: Innovation in policymaking often comes from learning. This learning extends beyond changes in policy outputs to include_x000D_ changes in the basic beliefs guiding policy formulation. Learning opportunities are provided by:_x000D_ · learning from past experiences (effective monitoring and evaluation)_x000D_ · observation and knowledge exchange (good practices, international cooperation)_x000D_ · consultancy (academic experts and practitioners)_x000D_ Flexibility refers to a government’s ability to adapt to and take advantage of developmental opportunities inherent to_x000D_ a given political situation. Flexibility and learning allow governments to replace failed policies with innovative ones._x000D_ If possible, provide empirical evidence on whether policy learning happens coincidentally or if there are institutionalized_x000D_ mechanisms that facilitate innovation and flexibility in policymaking._x000D_ Please consider if innovation and flexibility correspond with the normative framework of the BTI in terms of striving_x000D_ for democracy and a market economy. The maximum score for autocracies is 5 points.
  \item \strong{bs_bti_q15_1}: To what extent does the government make efficient use of available human, financial and organizational resources?
  \item \strong{bs_bti_q15_2}: As many policies have conflicting objectives, reflect competing political interests and affect other policies, the government_x000D_ has to ensure that its overall policy is coherent._x000D_ Successful coordination should:_x000D_ · assure that trade-offs between policy goals are well balanced_x000D_ · introduce horizontal forms of coordination to mediate between different departments of the state administration_x000D_ · ascribe responsibilities in a transparent manner to avoid the negligence of tasks, redundancies or friction between_x000D_ different government branches._x000D_ Various coordination styles— hierarchic-bureaucratic, informal-network, personalist, centralized, decentralized_x000D_ etc. — are possible and may be functionally equivalent. What matters is their impact on policy coherence
  \item \strong{bs_bti_q2_1}: This indicator assesses if: general elections are regularly conducted on the national level; universal suffrage with secret ballot is ensured; several parties with different platforms are able to run; and if political posts are filled according to election outcome. For this indicator evaluated the quality of elections considering if; the electoral management body is impartial and effective; registration procedures for voters, candidates and parties are transparent and fair; the polling procedures, including vote count, results verification and complaint resolution, are conducted in a transparent, impartial and correct manner; fair and equal media access is ensured for all candidates and parties; polling is accessible, secure and secret to ensure effective participation. This indicator is graded on a 1 to 10 scale, and a score over 6 is considered free and fair elections
  \item \strong{bs_bti_q2_3}: This indicator assesses to what extent can individuals form and join independent political or civic groups? To what extent can these groups operate and assemble freely? This considers: if the constitution guarantees freedom of association and assembly, and if these laws are enforced; if there are severe restrictions on assembly and association for all citizens or for particular groups; if the government uses intimidation, harassment or threats of retaliation to prevent citizens from exercising the rights to association and assembly (e.g. by arbitrarily arresting, detaining and imprWorld Bankning peaceful demonstrators or using excessive force); if the government uses transparent and non-discriminatory criteria in evaluating requests for permits to associate and/or assemble; if groups are able to operate free from unwarranted state intrusion or interference in their affairs
  \item \strong{bs_bti_q3_1}: To what extent is there a working separation of powers (checks and balances)?
  \item \strong{bs_bti_q3_2}: An independent judiciary has the ability and autonomy to: interpret and review existing laws, legislation and policies, both public and civil, pursue its own reasoning, free from the influence of political decision-makers or powerful groups and individuals and from corruption, develop a differentiated organization, including legal education, jurisprudence, regulated appointment of the judiciary, rational proceedings, professionalism, channels of appeal and court administration
  \item \strong{bs_bti_q7_2}: To what level have the fundamentals of market-based competition developed? The main focus of this question is to assess if free and fair competition is guaranteed by an institutional framework that ensures unrestricted participation in the market and a level playing field for all market participants. From "Market competition is present only in small segments of the economy and its institutional framework is rudimentary. Rules for market participants are unreliable and frequently set arbitrarily. The informal sector is large." to "Market competition is consistently defined and implemented both macroeconomically and microeconomically. There are state-guaranteed rules for market competition with equal opportunities for all market participants. The informal sector is very small."
  \item \strong{bs_bti_q7_4}: International standards of banking systems are defined by the Basel Accords (www.bis.org). These standards require_x000D_ banks to hold a minimum share of capital equity in relation to their assets and to the risk the bank exposes itself to_x000D_ through its lending and investment practices. They also demand that banks undergo a supervWorld Bankry review process,_x000D_ and that they disclose information about their economic activities._x000D_ A solid capital market is characterized by a past record of low shares of nonperforming loans, prudential requirements_x000D_ for bank transactions in foreign currency, enforcement of rules on disclosure, independence of financial regulators_x000D_ and hard budget constraints between companies, banks and the public sector._x000D_ Note: Please make sure you provide information on capital-adequacy ratios, disclosure rules, supervision, share of_x000D_ non-performing loans and hard budget constraints. You should also elaborate on reforms in the financial sector._x000D_ Quantitative Reference Indicators: Bank capital to assets ratio · Bank nonperforming loans
  \item \strong{bs_bti_q8_1}: To what extent does the monetary authority pursue and communicate a consistent monetary stabilization policy?
  \item \strong{bs_bti_q8_2}: To what extent do the government’s budgetary policies support fiscal stability?
  \item \strong{bs_bti_si}: The democracy status is made up of five criterion: Stateness, measured by the normative statement "There is clarity about the nation's existence as a state with adequately established and differentiated power structures"; Political Participation, measured by the normative statement "The populace decides who rules and it has other political freedoms"; Rule of Law measured by the normative statement "State powers check and balance one another and ensure civil rights"; Stability of Democratic Institutions, measured by the normative statement "Democratic Institutions are capable of performing, and they are adequately accepted"; and Political and Social Integration, measured by the normative statement "Stable patterns of representation exist for mediating between society and the state; there is also a consolidated civic culture"
  \item \strong{imf_world_rg_rm_gdp}: Total Grants in Percent of GDP
  \item \strong{imf_world_rs_rm_gdp}: Social Contributions in Percent of GDP
  \item \strong{imf_world_rt_rm_gdp}: Captures compulsory transfers to the central government for public purposes. Certain compulsory transfers such as fines, penalties, and most social security contributions are excluded. Refunds and corrections of erroneously collected tax revenue are treated as negative revenue.
  \item \strong{imf_world_rtgse_rm_gdp}: Excise Tax Revenue in Percent of GDP
  \item \strong{imf_world_rtgsgv_rm_gdp}: VAT Revenue in Percent of GDP :Some countries refer to VAT as the tax as General Sales Tax (GST) or Goods and Services Tax (GST)
  \item \strong{imf_world_rti_rm_gdp}: Income Tax Revenue in Percent of GDP
  \item \strong{imf_world_rtic_rm_gdp}: Corporate Income Tax Revenue in Percent of GDP
  \item \strong{imf_world_rtii_rm_gdp}: Individual Income Tax Revenue in Percent of GDP
  \item \strong{imf_world_rtp_rm_gdp}: Property Tax Revenue in Percent of GDP
  \item \strong{imf_world_rtpay_rm_gdp}: Taxes on Payroll and Workforce Revenue in Percent of GDP
  \item \strong{imf_world_rtt_rm_gdp}: Trade Tax Revenue in Percent of GDP.Gross domestic expenditures on research and development (R&D), expressed as a percent of GDP. They include both capital and current expenditures in the four main sectors: Business enterprise (tade), Government, Higher education and Private non-profit. R&D covers basic research, applied research, and experimental development.
  \item \strong{wb_girg_6}: It captures how policymakers interact with stakeholders when shaping regulations affecting business communities. It considers: publication of forward regulatory plans; consultation on proposed regulations; report back on the results of that consultation process; conduct regulatory impact assessment; laws are made publicly accessible.
  \item \strong{wb_gtmi_cgsi}: The Core Government Systems Index (17 indicators) captures the key aspects of a whole-of-government approach, including government cloud, interoperability framework and other platforms.
  \item \strong{wb_gtmi_dcei}: The Digital Citizen Engagement Index (6 indicators) measures aspects of public participation platforms, citizen feedback mechanisms, open data, and open government portals.
  \item \strong{wb_gtmi_gtei}: The GovTech Enablers Index (16 indicators) captures strategy, Institutions, laws, and regulations, as well as digital skills, and innovation policies and programs, to foster GovTech.
  \item \strong{wb_gtmi_i_12}: Is there an e-Procurement portal? Measured by 0= No, 1= Implementation in progress, 2= Yes (in use)
  \item \strong{wb_gtmi_i_13}: Is there a Debt Management System (DMS) in place? (foreign and domestic debt). This indicator is measured with the scale 0= No, 1= Implementation in progress, 2= Yes (in use)
  \item \strong{wb_gtmi_i_14}: Is there a Public Investment Management System (PIMS) in place? Measured by 0= No, 1= Implementation in progress, 2= Yes (in use)
  \item \strong{wb_gtmi_i_5}: Is there an operational FMIS in place to support core PFM functions? This is measured on the scale 0= No/Unknown, 1= Implementation in progress, 2= Yes (in use)
  \item \strong{wb_gtmi_i_6}: Is there a TSA supported by FMIS to automate payments and bank reconciliation? Measured with the scale 0= No, 1= Implementation in progress, 2= Yes (in use)
  \item \strong{wb_gtmi_i_7}: Is there a Tax Management Information System in place? Measured with scale 0= No, 1= Implementation in progress, 2= Yes (in use)
  \item \strong{wb_gtmi_i_8}: Is there a Customs Management Information System in place? Measured by 0= No, 1= Implementation in progress, 2= Yes (in use)
  \item \strong{wb_gtmi_psdi}: The Public Service Delivery Index (9 indicators) measures the maturity of online public service portals, with a focus on citizen centric design and universal accessibility.
  \item \strong{wb_wdi_dt_dod_mwbg_cd}: IBRD loans and IDA credits are public and publicly guaranteed debt extended by the World Bank Group. The International Bank for Reconstruction and Development (IBRD) lends at market rates. Credits from the International Development Association (IDA) are at concessional rates. Data are in current U.S. dollars.
  \item \strong{wb_wdi_dt_nfl_pcbo_cd}: Commercial bank and other lending includes net commercial bank lending (public and publicly guaranteed and private nonguaranteed) and other private credits. Data are in current U.S. dollars.
  \item \strong{wb_wdi_gc_rev_xgrt_gd_zs}: Revenue is cash receipts from taxes, social contributions, and other revenues such as fines, fees, rent, and income from property or sales. Grants are also considered as revenue but are excluded here.
  \item \strong{wb_wdi_gc_tax_expt_zs}: Taxes on exports are all levies on goods being transported out of the country or services being delivered to nonresidents by residents. Rebates on exported goods that are repayments of previously paid general consumption taxes, excise taxes, or import duties are deducted from the gross amounts receivable from these taxes, not from amounts receivable from export taxes.
  \item \strong{wb_wdi_gc_tax_gsrv_rv_zs}: Taxes on goods and services include general sales and turnover or value added taxes, selective excises on goods, selective taxes on services, taxes on the use of goods or property, taxes on extraction and production of minerals, and profits of fiscal monopolies.
  \item \strong{wb_wdi_gc_tax_impt_zs}: Customs and other import duties are all levies collected on goods that are entering the country or services delivered by nonresidents to residents. They include levies imposed for revenue or protection purposes and determined on a specific or ad valorem basis as long as they are restricted to imported goods or services.
  \item \strong{wb_wdi_gc_tax_intt_rv_zs}: Taxes on international trade include import duties, export duties, profits of export or import monopolies, exchange profits, and exchange taxes.
  \item \strong{wb_wdi_gc_xpn_comp_zs}: Compensation of employees consists of all payments in cash, as well as in kind (such as food and housing), to employees in return for services rendered, and government contributions to social insurance schemes such as social security and pensions that provide benefits to employees.
  \item \strong{wb_wdi_gc_xpn_gsrv_zs}: Taxes on international trade include import duties, export duties, profits of export or import monopolies, exchange profits, and exchange taxes.
  \item \strong{wb_wdi_gc_xpn_trft_zs}: Goods and services include all government payments in exchange for goods and services used for the production of market and nonmarket goods and services. Own-account capital formation is excluded.
  \item \strong{wb_wdi_ic_bus_ndns_zs}: New businesses registered are the number of new limited liability corporations (or its equivalent) registered in the calendar year.
  \item \strong{wb_wdi_ic_bus_nreg}: New businesses registered are the number of new limited liability corporations (or its equivalent) registered in the calendar year.
  \item \strong{wb_wdi_ie_ppi_watr_cd}: Investment in water and sanitation projects with private participation refers to commitments to  infrastructure projects in water and sanitation that have reached financial closure and directly or indirectly serve the public. Movable assets, incinerators, standalone solid waste projects, and small projects are excluded. The types of projects included are management and lease contracts, operations and management contracts with major capital expenditure, greenfield projects (in which a private entity or a public-private joint venture builds and operates a new facility), and divestitures. Investment commitments are the sum of investments in facilities and investments in government assets. Investments in facilities are the resources the project company commits to invest during the contract period either in new facilities or in expansion and modernization of existing facilities. Investments in government assets are the resources the project company spends on acquiring government assets such as state-owned enterprises, rights to provide services in a specific area, or the use of specific radio spectrums. Data is presented based on investment year. Data are in current U.S. dollars.
  \item \strong{wb_wdi_iq_spi_ovrl}: The SPI overall score is a composite score measuing country performance across five pillars: data use, data services, data products, data sources, and data infrastructure.  The new Statistical Performance Indicators (SPI) will replace the Statistical Capacity Index (SCI), which the World Bank has regularly published since 2004. Although the goals are the same, to offer a better tool to measure the statistical systems of countries, the new SPI framework has expanded into new areas including in the areas of data use, administrative data, geospatial data, data services, and data infrastructure. The SPI provides a framework that can help countries measure where they stand in several dimensions and offers an ambitious measurement agenda for the international community.
  \item \strong{wb_wdi_iq_spi_pil1}: The data use overall score is a composite score measuring the demand side of the statistical system.  The data use  pillar is segmented by five types of users: (i) the legislature, (ii) the executive branch, (iii) civil society (including sub-national actors), (iv) academia and (v) international bodies.  Each dimension would have associated indicators to measure performance. A mature system would score well across all dimensions whereas a less mature one would have weaker scores along certain dimensions. The gaps would give insights into prioritization among user groups and help answer questions as to why the existing services are not resulting in higher use of national statistics in a particular segment.  Currently, the SPI only features indicators for one of the five dimensions of data use, which is data use by international organizations. Indicators on whether statistical systems are providing useful data to their national governments (legislature and executive branches), to civil society, and to academia are absent.  Thus the dashboard does not yet assess if national statistical systems are meeting the data needs of a large swathe of users.
  \item \strong{wb_wdi_iq_spi_pil2}: The data services pillar overall score is a composite indicator based on four dimensions of data services: (i) the quality of data releases, (ii) the richness and openness of online access, (iii) the effectiveness of advWorld Bankry and analytical services related to statistics, and (iv) the availability and use of data access services such as secure microdata access. AdvWorld Bankry and analytical services might incorporate elements related to data stewardship services including input to national data strategies, advice on data ethics and calling out misuse of data in accordance with the Fundamental Principles of Official Statistics.
  \item \strong{wb_wdi_iq_spi_pil3}: The data products overall score is a composite score measureing whether the country is able to produce relevant indicators, primarily related to SDGs.  The data products (internal process) pillar is segmented by four topics and organized into (i) social, (ii) economic, (iii) environmental, and (iv) institutional dimensions using the typology of the Sustainable Development Goals (SDGs). This approach anchors the national statistical system’s performance around the essential data required to support the achievement of the 2030 global goals, and enables comparWorld Bankns across countries so that a global view can be generated while enabling country specific emphasis to reflect the user needs of that country.
  \item \strong{wb_wdi_iq_spi_pil4}: The data sources overall score is a composity measure of whether countries have data available from the following sources: Censuses and surveys, administrative data, geospatial data, and private sector/citizen generated data.  The data sources (input) pillar is segmented by four types of sources generated by (i) the statistical office (censuses and surveys), and sources accessed from elsewhere such as (ii)  administrative data, (iii) geospatial data, and (iv) private sector data and citizen generated data. The appropriate balance between these source types will vary depending on a country’s institutional setting and the maturity of its statistical system. High scores should reflect the extent to which the sources being utilized enable the necessary statistical indicators to be generated. For example, a low score on environment statistics (in the data production pillar) may reflect a lack of use of (and low score for) geospatial data (in the data sources pillar). This type of linkage is inherent in the data cycle approach and can help highlight areas for investment required if country needs are to be met.
  \item \strong{wb_wdi_iq_spi_pil5}: The data infrastructure  pillar  overall score measures the hard and soft infrastructure segments, itemizing essential cross cutting requirements for an effective statistical system.  The segments are: (i) legislation and governance covering the existence of laws and a functioning institutional framework for the statistical system; (ii) standards and methods addressing compliance with recognized frameworks and concepts; (iii) skills including level of skills within the statistical system and among users (statistical literacy); (iv) partnerships reflecting the need for the statistical system to be inclusive and coherent; and (v) finance mobilized both domestically and from donors.
  \item \strong{wb_wdi_tm_val_fuel_zs_un}: Fuels comprise the commodities in SITC section 3 (mineral fuels, lubricants and related materials).
  \item \strong{wb_gtmi_pfm_mis}: CLIAR made Index of PFM MIS based on GTMI data. Considers whether there's an operational FMIS (GTMI_I-5), a TSA (GTMI_I-6), a Tax MIS (GTMI_I-7), a DMS (GTMI_I-13), Customs MIS (GTMI_I-8), a PIMS (GTMI_I-14), and an e-Procurement system in place.
}
}
\examples{
data(d360_efi_data)
str(d360_efi_data)
head(d360_efi_data)
}
\keyword{data360}
\keyword{datasets}
\keyword{efi}
\keyword{governance}
\keyword{institutions}
