% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{debt_transparency}
\alias{debt_transparency}
\title{Debt Reporting Heat Map}
\format{
A data frame with 373 rows and 13 variables:
\describe{
  \item{\code{country_code}}{character World Bank country code}
  \item{\code{year}}{integer Year}
  \item{\code{wb_debt_transp_data}}{double. Data accessibility}
  \item{\code{wb_debt_transp_instrument}}{double. Instrument coverage}
  \item{\code{wb_debt_transp_sectorial}}{double. Sectoral coverage}
  \item{\code{wb_debt_transp_information}}{double. Information on recently contracted loans}
  \item{\code{wb_debt_transp_periodicity}}{double. Periodicity of reporting}
  \item{\code{wb_debt_transp_time}}{double. Timeliness (time lag) of reporting}
  \item{\code{wb_debt_transp_dms}}{double. Availability of a debt management strategy}
  \item{\code{wb_debt_transp_abp}}{double. Availability of an annual borrowing plan}
  \item{\code{wb_debt_transp_addition}}{double. Other debt statistics (e.g., contingent liabilities)}
  \item{\code{wb_debt_transp_index}}{double. Composite index: row average of all debt transparency indicators}
}
}
\source{
https://www.worldbank.org/en/topic/debt/brief/debt-transparency-report/
}
\usage{
debt_transparency
}
\description{
How transparent are IDA countries in their debt reporting practices? This heat map presents an assessment based on the availability, completeness, and timeliness of public debt statistics and debt management documents posted on national authorities' websites. The assessment will be updated annually.
}
\keyword{datasets}
