% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_data360_funs.R
\name{extract_dataset_id}
\alias{extract_dataset_id}
\title{Extract Dataset ID from Indicator Name}
\usage{
extract_dataset_id(input_id, splitchar = ".")
}
\arguments{
\item{input_id}{A character string representing the full indicator name.}

\item{splitchar}{A character string used as the delimiter to split the input. Default is \code{'.'}.}
}
\value{
A character string representing the extracted dataset ID. If the delimiter
        is not found twice, the original \code{input_id} is returned.
}
\description{
This function extracts the EFI source (dataset ID) required for an API pull
from a given indicator name. It identifies and returns the substring between
the first and second occurrences of a specified delimiter character.
}
\examples{
extract_dataset_id("WDI.GDP.PC")
extract_dataset_id("IMF.WEO.GDP", splitchar = ".")

}
