% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{family_order}
\alias{family_order}
\title{Family Order and Names}
\format{
A tibble with 13 rows and 2 variables:
\describe{
  \item{family_order}{Numeric code (integer) indicating the institutional family order.}
  \item{family_name}{Character string with the institutional family name.}
}
}
\source{
Created manually for internal project use.
}
\usage{
family_order
}
\description{
A tibble mapping numeric family order codes to their corresponding
institutional family names. This dataset is intended for use in
organizing or categorizing indicators in dashboards, reports, or analyses.
}
\keyword{datasets}
