% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_funs.R
\name{flag_continued}
\alias{flag_continued}
\title{Flag continuity of an indicator series}
\usage{
flag_continued(indicator, year_id, ref_year)
}
\arguments{
\item{indicator}{A vector of indicator values. May include `NA`s.}

\item{year_id}{A vector of years corresponding to the indicator values.}

\item{ref_year}{A numeric value indicating the most recent year of reference.}
}
\value{
An integer flag:
\itemize{
  \item `1` if at least one non-missing value is available in the last five years.
  \item `0` otherwise.
}
}
\description{
Returns a flag indicating whether an indicator has been updated
within the last five years relative to a reference year.
}
