% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_funs.R
\name{flag_country}
\alias{flag_country}
\title{Flag sufficient country coverage}
\usage{
flag_country(indicator, country_id, year_id, ref_year, country_region_list)
}
\arguments{
\item{indicator}{A vector of indicator values.}

\item{country_id}{A vector of country identifiers.}

\item{year_id}{A vector of years.}

\item{ref_year}{A numeric value indicating the most recent year of reference.}

\item{country_region_list}{A data frame mapping `country_code` to `region`.}
}
\value{
An integer flag (`1` or `0`).
}
\description{
Returns a flag indicating whether an indicator achieves a minimum
level of country and regional coverage in the last five years.
}
\details{
The flag is set to `1` if either:
\itemize{
  \item At least 100 countries have non-missing values in the last five years; or
  \item At least 50 countries have non-missing values and all 7 regions are represented.
}

Otherwise, the flag is set to `0`.
}
