% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_funs.R
\name{flag_minimum_coverage}
\alias{flag_minimum_coverage}
\title{Flag minimum coverage of an indicator}
\usage{
flag_minimum_coverage(indicator, country_id, year_id)
}
\arguments{
\item{indicator}{A vector of indicator values.}

\item{country_id}{A vector of country identifiers.}

\item{year_id}{A vector of years.}
}
\value{
An integer flag (`1` or `0`)
}
\description{
Returns a flag indicating whether an indicator has sufficient
country coverage over time.
}
\details{
The flag is set to `1` if the indicator has values in at least two years
where each of those years covers 10 or more distinct countries.
Otherwise, the flag is set to `0`.
}
