% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_management_funs.R
\name{flag_mismatched_indicators}
\alias{flag_mismatched_indicators}
\title{Flag Mismatched Indicators to Dictionary}
\usage{
flag_mismatched_indicators(df, dictionary)
}
\arguments{
\item{df}{A data frame containing extracted indicators along with identifying columns such as `country_code`, `country_name`, and `year`.}

\item{dictionary}{A data frame containing a column named `variable` with the list of valid variable names.}
}
\value{
A named list containing a tibble with one column, `variable`, listing the mismatched variables found in `df` but not in `db_variables$variable`.
The name of the element in the list is generated by appending `_mismatched_vars` to the original name of the `df` object.
}
\description{
Identifies variable names in a data frame that are not included in a reference dictionary of valid indicators `db_variables`.
}
