% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_management.R
\name{flag_missing_indicators}
\alias{flag_missing_indicators}
\title{Flag Missing Indicators Compared to Dictionary}
\usage{
flag_missing_indicators(db_variables, df, source_type = NULL)
}
\arguments{
\item{db_variables}{A data frame containing a column named `variable` with the list of valid variable names, and optionally a `source` column.}

\item{df}{A data frame containing extracted indicators along with identifying columns such as `country_code`, `country_name`, and `year`.}

\item{source_type}{Optional. A character string specifying the source to filter the dictionary (e.g., "WDI").}
}
\value{
A named list containing a tibble with one column, `variable`, listing the variables from `db_variables` that are missing in `df`.
The name of the element in the list is generated by appending `_missing_vars` to the original name of the `df` object.
}
\description{
Identifies variables that are in the dictionary (`db_variables`) but missing from the given data frame.
}
