% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{romelli}
\alias{romelli}
\title{Central Bank Independence Index (Romelli Update of Cukierman et al.)}
\format{
A tibble with 8,128 rows and 3 variables:
\describe{
  \item{country_code}{A 3-letter World Bank country code (e.g., "USA", "FRA", "ARG").}
  \item{year}{The calendar year (numeric).}
  \item{romelli_cbi_central_bank_independence}{Cukierman central bank independence index (scaled 0 to 1).}
}
}
\source{
Romelli, Davide (2022). \emph{The Political Economy of Reforms in Central Bank Design:
Evidence from a New Dataset}. Review of International Organizations.
\url{https://cbidata.org/}
}
\usage{
romelli
}
\description{
based on the unweighted index developed by Cukierman et al. (1992) and recomputed by Romelli using
updated legal and institutional data.
}
\details{
The index ranges from 0 to 1, where 0 represents the lowest level of independence and 1 the highest.
It measures de jure independence, focusing on aspects such as governor appointment, monetary policy formulation,
central bank objectives, and limitations on lending to the government.
}
\references{
Cukierman, A., Webb, S. B., & Neyapti, B. (1992). \emph{Measuring the Independence of Central Banks
and Its Effect on Policy Outcomes}. The World Bank Economic Review, 6(3), 353–398.
}
\keyword{datasets}
