% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_management.R
\name{scale_values}
\alias{scale_values}
\title{Scale Values to [0, 1] Range}
\usage{
scale_values(x)
}
\arguments{
\item{x}{A numeric vector to be scaled.}
}
\value{
A numeric vector where all non-missing values are
rescaled to the range [0, 1]. If all elements of `x` are `NA`,
the result will be `NA`.
}
\description{
This function rescales a numeric vector to a 0–1 range using
min-max normalization. Missing values (`NA`) are ignored
when computing the minimum and maximum.
}
\examples{
scale_values(c(1, 2, 3, 4, 5))
# Returns: 0.00 0.25 0.50 0.75 1.00

scale_values(c(10, 20, NA, 30))
# Returns: 0.00 0.50   NA  1.00

}
