% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{closeness_to_frontier_dynamic}
\alias{closeness_to_frontier_dynamic}
\title{Closeness to Frontier (CTF) Scores, Dynamic}
\format{
A data frame with 2592 rows and 116 variables:
\describe{
  \item{\code{country_code}}{character Three-letter World Bank country code.}
  \item{\code{country_name}}{character World Bank country name.}
  \item{\code{income_group}}{character COLUMN_DESCRIPTION}
  \item{\code{region}}{character World Bank region.}
  \item{\code{year}}{double Year.}
}
}
\usage{
closeness_to_frontier_dynamic
}
\description{
Closeness to frontier scores computed in a dynamic time frame: every two years between 2014 and 2023. For a complete list of variable descriptions, please refer to the `db_variables` data frame.
}
\details{
For additional details on how the closeness to frontier scores are computed, please refer to the CLIAR Methodology Note.
}
\keyword{datasets}
