% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{debt_transparency}
\alias{debt_transparency}
\title{Debt Reporting Heat Map}
\format{
A data frame with 373 rows and 13 variables:
\describe{
  \item{\code{country_code}}{character World Bank country code}
  \item{\code{year}}{integer Year}
  \item{\code{wb_debt_transp_index}}{double. Composite index: row average of all debt transparency indicators}
}
#' @details
Below are the longer descriptions for each variable (when provided):
\itemize{
  \item \strong{wb_debt_transp_index}: The debt transparency index is obtained as a simple average of the World Bank's debt reporting heatmap sub-components (Debt Transparency in Developing Economies) that measure instrument and sectorial coverage of debt statistics, availability of financial terms on new loans, and publication of DMS and ABP.
}
}
\source{
https://www.worldbank.org/en/topic/debt/brief/debt-transparency-report/
}
\usage{
debt_transparency
}
\description{
How transparent are IDA countries in their debt reporting practices? This heat map presents an assessment based on the availability, completeness, and timeliness of public debt statistics and debt management documents posted on national authorities' websites. The assessment will be updated annually.
}
\keyword{datasets}
